requirements.txt
fastapi
uvicorn
numpy
Pillow
tensorflow
scikit-learn
jinja2
